flink_consumer/requirements.txt
apache-flink
vaderSentiment
python-dotenv
# safetensors
# transformers